# packages/shared-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Raumdeuter API",
    "version": "0.0.1"
  },
  "paths": {
...
    "/teams/minutes-distribution": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Minutes Distribution",
        "operationId": "minutes_distribution_teams_minutes_distribution_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Season start year (e.g. 2025 for 25-26)",
              "title": "Season"
            },
            "description": "Season start year (e.g. 2025 for 25-26)"
          },
          {
            "name": "club",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Club name as stored in parquet `club`",
              "title": "Club"
            },
            "description": "Club name as stored in parquet `club`"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MinutesDistributionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/minutes-distribution/league": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Minutes Distribution League",
        "operationId": "minutes_distribution_league_teams_minutes_distribution_league_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Season start year (e.g. 2025 for 25-26)",
              "title": "Season"
            },
            "description": "Season start year (e.g. 2025 for 25-26)"
          },
          {
            "name": "league",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "League name as stored in parquet `league`",
              "title": "League"
            },
            "description": "League name as stored in parquet `league`"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueMinutesOverviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/squad-value/league": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Squad Value League",
        "operationId": "squad_value_league_teams_squad_value_league_get",
        "parameters": [
          {
            "name": "season",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Season start year, e.g. 2024 for 24-25",
              "title": "Season"
            },
            "description": "Season start year, e.g. 2024 for 24-25"
          },
          {
            "name": "league",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "League name as stored in parquet `league`",
              "title": "League"
            },
            "description": "League name as stored in parquet `league`"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SquadValueLeagueResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/teams/squad-value/history": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Squad Value History",
        "operationId": "squad_value_history_teams_squad_value_history_get",
        "parameters": [
          {
            "name": "club",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Club name as stored in parquet `club`",
              "title": "Club"
            },
            "description": "Club name as stored in parquet `club`"
          },
          {
            "name": "seasons",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "integer"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional explicit list of seasons; defaults to all loaded seasons.",
              "title": "Seasons"
            },
            "description": "Optional explicit list of seasons; defaults to all loaded seasons."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SquadValueHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeagueClubBand": {
        "properties": {
          "club": {
            "type": "string",
            "title": "Club"
          },
          "club_logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club Logo"
          },
          "total_minutes": {
            "type": "integer",
            "title": "Total Minutes"
          },
          "zone_shares": {
            "$ref": "#/components/schemas/ZoneShares"
          }
        },
        "type": "object",
        "required": [
          "club",
          "total_minutes",
          "zone_shares"
        ],
        "title": "LeagueClubBand"
      },
      "LeagueMinutesOverviewResponse": {
        "properties": {
          "league": {
            "type": "string",
            "title": "League"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "max_league_games": {
            "type": "integer",
            "title": "Max League Games"
          },
          "max_league_minutes": {
            "type": "integer",
            "title": "Max League Minutes"
          },
          "clubs": {
            "items": {
              "$ref": "#/components/schemas/LeagueClubBand"
            },
            "type": "array",
            "title": "Clubs"
          }
        },
        "type": "object",
        "required": [
          "league",
          "season",
          "max_league_games",
          "max_league_minutes",
          "clubs"
        ],
        "title": "LeagueMinutesOverviewResponse",
        "description": "League-wide minutes overview: one row per club, with each club's squad\nminutes broken down by age band (% of that club's total minutes). Sorted by\nyouth share descending, tie-broken alphabetically by club."
      },
...
      "MinutesDistributionPlayer": {
        "properties": {
          "wyscout_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wyscout Id"
          },
          "player": {
            "type": "string",
            "title": "Player"
          },
          "player_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Player Image Url"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "minutes": {
            "type": "integer",
            "title": "Minutes"
          },
          "matches": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matches"
          },
          "league_minutes_pct": {
            "type": "number",
            "title": "League Minutes Pct"
          },
          "age_zone": {
            "type": "string",
            "enum": [
              "youth",
              "peak",
              "experienced",
              "veteran"
            ],
            "title": "Age Zone"
          }
        },
        "type": "object",
        "required": [
          "player",
          "minutes",
          "league_minutes_pct",
          "age_zone"
        ],
        "title": "MinutesDistributionPlayer"
      },
      "MinutesDistributionResponse": {
        "properties": {
          "club": {
            "type": "string",
            "title": "Club"
          },
          "club_logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club Logo"
          },
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "max_league_games": {
            "type": "integer",
            "title": "Max League Games"
          },
          "max_league_minutes": {
            "type": "integer",
            "title": "Max League Minutes"
          },
          "zone_shares": {
            "$ref": "#/components/schemas/ZoneShares"
          },
          "players": {
            "items": {
              "$ref": "#/components/schemas/MinutesDistributionPlayer"
            },
            "type": "array",
            "title": "Players"
          }
        },
        "type": "object",
        "required": [
          "club",
          "season",
          "max_league_games",
          "max_league_minutes",
          "zone_shares",
          "players"
        ],
        "title": "MinutesDistributionResponse",
        "description": "Squad minutes distribution for a single club in a single season.\n\n`max_league_minutes = max_league_games * 90` from the static\n`LEAGUE_MAX_GAMES` table in `core/config.py`. Multi-stage / playoff leagues\nuse the regular-season game count; downstream callers clamp the per-player\npct to 100 in case a player accrues playoff minutes on top.\n\nAge bands are fixed: youth (<23), peak (<29), experienced (<34),\nveteran (>=34). `zone_shares` are % of the squad's total minutes played."
      },
...
      "SquadValueHistoryResponse": {
        "properties": {
          "club": {
            "type": "string",
            "title": "Club"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/SquadValueHistoryRow"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "club",
          "rows"
        ],
        "title": "SquadValueHistoryResponse"
      },
...
      "SquadValueLeagueResponse": {
        "properties": {
          "season": {
            "type": "integer",
            "title": "Season"
          },
          "league": {
            "type": "string",
            "title": "League"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/SquadValueTeamRow"
            },
            "type": "array",
            "title": "Teams"
          }
        },
        "type": "object",
        "required": [
          "season",
          "league",
          "teams"
        ],
        "title": "SquadValueLeagueResponse"
      },
      "SquadValueTeamRow": {
        "properties": {
          "club": {
            "type": "string",
            "title": "Club"
          },
          "club_logo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Club Logo"
          },
          "league": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "League"
          },
          "n_players": {
            "type": "integer",
            "title": "N Players"
          },
          "avg_age": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Age"
          },
          "total_xtv_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Xtv Eur"
          },
          "avg_xtv_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Xtv Eur"
          },
          "total_market_value_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Market Value Eur"
          },
          "avg_market_value_eur": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Market Value Eur"
          },
          "foreign_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Foreign Share",
            "description": "Share (0-1) of players whose Passport country differs from the modal squad passport."
          }
        },
        "type": "object",
        "required": [
          "club",
          "n_players"
        ],
        "title": "SquadValueTeamRow",
        "description": "Squad-level aggregates for one club in one season."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ZoneShares": {
        "properties": {
          "youth": {
            "type": "number",
            "title": "Youth",
            "default": 0.0
          },
          "peak": {
            "type": "number",
            "title": "Peak",
            "default": 0.0
          },
          "experienced": {
            "type": "number",
            "title": "Experienced",
            "default": 0.0
          },
          "veteran": {
            "type": "number",
            "title": "Veteran",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "ZoneShares",
        "description": "Share of total team minutes per age band, rounded to 1dp (percent)."
      }
    }
  }
}